+++ Suspension/src/Suspension.DataContracts/RunAnalysis/Dynamic/RunDynamicAnalysisResponseData.cs
@@ -1,4 +1,5 @@
 ﻿using MudRunner.Commons.DataContracts.Operation;
+using System.Collections.Generic;
 
 namespace MudRunner.Suspension.DataContracts.RunAnalysis.Dynamic
 {
@@ -8,13 +9,27 @@ namespace MudRunner.Suspension.DataContracts.RunAnalysis.Dynamic
     public class RunDynamicAnalysisResponseData : OperationResponseData
     {
         /// <summary>
-        /// The full name of solution file.
+        /// Class constructor.
         /// </summary>
-        public string FullFileName { get; set; }
+        public RunDynamicAnalysisResponseData()
+        {
+            this.FullFileNames = new List<string>();
+        }
 
         /// <summary>
-        /// The maximum results for analysis.
+        /// The full name of solution file.
+        /// </summary>
+        // TODO: Disponibilizar os arquivos para download.
+        public List<string> FullFileNames { get; set; }
+
+        /// <summary>
+        /// The maximum result for analysis.
         /// </summary>
         public DynamicAnalysisResult MaximumResult { get; set; }
+
+        /// <summary>
+        /// The maximum deformation result.
+        /// </summary>
+        public DynamicAnalysisResult MaximumDeformationResult { get; set; }
     }
 }