+++ Suspension/CHANGELOG.md
@@ -7,12 +7,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## UNRELEASED
 ### Fixed
  - RunDynamicAnalysis operation that was not building correctly the equivalent force vector.
+### Changed
+ - RunDynamicAnalysis operations to calculate the system deformation.
 ### Added
  - Enum CurveType.
  - Class BaseExcitation.
  - Class DynamicAnalysisResult.
  - Properties 'ConsiderLargeDisplacements', 'BaseExcitation' and 'CarSpeed' on class RunDynamicAnalysisRequest.
- - Properties 'MaximumResult' and 'DynamicAnalysisResult' on class RunDynamicAnalysisResponseData.
+ - Properties 'MaximumResult' and 'MaximumDeformationResult' on class RunDynamicAnalysisResponseData.
  - Operations RunAmplitudeDynamicAnalysis, RunHalfCarSixDofDynamicAnalysis and RunHalfCarSixDofAmplitudeDynamicAnalysis.
 
 ## [1.2.0] - 2022-01-13